// onbkp-next/src/services/database/types.ts
export interface Database {
    id_database: string;
    id_user: string
    database_description: string
    database_name: string
    database_type: string
    server_name: string
    host: string
    port: string
    created_at: string
    observation: string
    routines: string
    retentions: string
    status: string
    scope: string
}